requirements.txt
torch
numpy
pandas
matplotlib
tqdm
biopython
